# 2.5x- 3x faster copyFile on osx