* fix proposal 1 from https://github.com/nim-lang/RFCs/issues/330

## benchmarks
## example 1
download https://releases.ubuntu.com/16.04.7/ubuntu-16.04.6-desktop-i386.iso
```nim
when true:
  import std/[times, os, algorithm]
  proc main()=
    let file = "/Users/timothee/Downloads/ubuntu-16.04.6-desktop-i386.iso"
    let file2 = "/tmp/D20210130T183001"
    echo getFileSize(file)
    for i in 0..<3:
      let t = epochTime()
      copyFile(file, file2)
      let t2 = epochTime()
      let dt = t2 - t
      echo dt
  main()
```

on osx with -d:danger:
before PR:
1677721600
3.034269094467163
3.089644908905029
3.209890127182007

after PR:
1677721600
1.278779029846191
1.217450857162476
1.241599082946777


nim --eval:'echo 3.034269094467163/1.217450857162476'
2.492313407655042

(this could be added to `tests/benchmarks/tos_copyfile.nim` in future work)

## example 2: example with copyDir
`copyDir("/Users/timothee/git_clone/llvm-project/build/tools", "/tmp/d01")`
before PR: 40s
after PR: 13.8s => 2.9x speedup

```
du -sh /Users/timothee/git_clone/llvm-project/build/tools
7.5G    /Users/timothee/git_clone/llvm-project/build/tools
find /Users/timothee/git_clone/llvm-project/build/tools | wc -l
    3798
```

## future work
* investigate why shell command `cp` is still faster; perhaps multithreading is used? but note that the speedup also occurs for single files.
  * example 1 gives: 0.00s user 0.78s system 93% cpu 0.846 total (1.4X faster than this PR)

  * example 2 gives: 0.09s user 5.96s system 51% cpu 11.690 total (1.18x faster than this PR)

(the disk is "hot", ie we don't measure cold start here; i'm using macbook pro 64 GB 2667 MHz DDR4, 2TB flash storage in all experiments)